separate BLOG into another app
