+++ catalog/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from catalog.models import Category, Product, BlogArticle, Version
+from catalog.models import Category, Product, Version
 
 
 @admin.register(Category)
@@ -14,12 +14,6 @@ class CategoryAdmin(admin.ModelAdmin):
     list_display = ('pk', 'name', 'price', 'category',)
     list_filter = ('category',)
     search_fields = ('name', 'description',)
-
-
-@admin.register(BlogArticle)
-class BlogArticleAdmin(admin.ModelAdmin):
-    list_display = ('title', 'views_counter',)
-    search_fields = ('title', 'content',)
 
 
 @admin.register(Version)
