+++ catalog/templates/catalog/includes/inc_menu.html
@@ -10,7 +10,7 @@
                 <h4>Навигация</h4>
                 <ul class="list-unstyled">
                     <li><a href="{% url 'catalog:product_list' %}" class="text-white">Главная</a></li>
-                    <li><a href="{% url 'catalog:blog_list' %}" class="text-white">Наш блог!</a></li>
+                    <li><a href="{% url 'blog:blog_list' %}" class="text-white">Наш блог!</a></li>
                     {% if user.is_authenticated %}
                     <li><a href="{% url 'users:logout' %}" class="text-white">Выйти</a></li>
                     {% else %}
