+++ catalog/views.py
@@ -5,7 +5,7 @@ from django.urls import reverse_lazy
 from django.views import View
 from django.views.generic import ListView, DetailView, CreateView, TemplateView, UpdateView, DeleteView
 from catalog.forms import ProductForm, VersionForm
-from catalog.models import Product, BlogArticle, Version
+from catalog.models import Product, Version
 
 
 class ProductListView(ListView):
@@ -65,34 +65,3 @@ class ProductUpdateView(LoginRequiredMixin, UpdateView):
 class ProductDeleteView(LoginRequiredMixin, DeleteView):
     model = Product
     success_url = reverse_lazy("catalog:product_list")
-
-
-class BlogArticleListView(ListView):
-    model = BlogArticle
-
-
-class BlogArticleDetailView(DetailView):
-    model = BlogArticle
-
-    def get_object(self, queryset=None):
-        self.object = super().get_object(queryset)
-        self.object.views_counter += 1
-        self.object.save()
-        return self.object
-
-
-class BlogArticleCreateView(CreateView):
-    model = BlogArticle
-    fields = ("title", "content", "preview_image", "is_published", "views_counter")
-    success_url = reverse_lazy("catalog:blog_list")
-
-
-class BlogArticleUpdateView(UpdateView):
-    model = BlogArticle
-    fields = ("title", "content", "preview_image", "is_published", "views_counter")
-    success_url = reverse_lazy("catalog:blog_list")
-
-
-class BlogArticleDeleteView(DeleteView):
-    model = BlogArticle
-    success_url = reverse_lazy("catalog:blog_list")
