+++ config/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
 
     'catalog',
     'users',
+    'blog',
 ]
 
 MIDDLEWARE = [
